+++ Application/Model/Product.php
@@ -16,9 +16,41 @@ class Product extends Product_parent
     /**
      * Method checks if product has variants.
      *
+     * @deprecated Use oeEcondaTrackingHasVariants method instead.
+     *
      * @return bool
      */
     public function oePersonalizationHasVariants()
+    {
+        return $this->oeEcondaTrackingHasVariants();
+    }
+
+    /**
+     * @deprecated Use oeEcondaTrackingGetSku method instead.
+     *
+     * @return string|null
+     */
+    public function oePersonalizationGetSku()
+    {
+        return $this->oeEcondaTrackingGetSku();
+    }
+
+    /**
+     * @deprecated Use oeEcondaTrackingGetProductId method instead.
+     *
+     * @return string
+     */
+    public function oePersonalizationGetProductId()
+    {
+        return $this->oeEcondaTrackingGetProductId();
+    }
+
+    /**
+     * Method checks if product has variants.
+     *
+     * @return bool
+     */
+    public function oeEcondaTrackingHasVariants()
     {
         $result = false;
         $sId = $this->getId();
@@ -36,9 +68,9 @@ class Product extends Product_parent
     }
 
     /**
-     * @return string|null
+     * @return null|string
      */
-    public function oePersonalizationGetSku()
+    public function oeEcondaTrackingGetSku()
     {
         $sku = null;
         if (!$this->oePersonalizationHasVariants()) {
@@ -51,7 +83,7 @@ class Product extends Product_parent
     /**
      * @return string
      */
-    public function oePersonalizationGetProductId()
+    public function oeEcondaTrackingGetProductId()
     {
         /** @var Article $parent */
         $parent = $this->getParentArticle();